+++ legacy/creator-subgraph/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @zoralabs/nft-creator-subgraph
 
+## 0.3.41
+
+### Patch Changes
+
+- Updated dependencies [262d1539]
+  - @zoralabs/protocol-deployments@0.6.2
+
+## 0.3.40
+
+### Patch Changes
+
+- Updated dependencies [e4938846]
+  - @zoralabs/comments-contracts@0.1.0
+
 ## 0.3.39
 
 ### Patch Changes