+++ nft-docs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # nft-docs
 
+## 0.0.28
+
+### Patch Changes
+
+- Updated dependencies [a60f441b]
+  - @zoralabs/coins-sdk@0.2.11
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [262d1539]
+  - @zoralabs/protocol-deployments@0.6.2
+  - @zoralabs/protocol-sdk@0.13.11
+  - @zoralabs/coins-sdk@0.2.10
+
 ## 0.0.26
 
 ### Patch Changes